requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
flatbuffers==25.12.19
fonttools==4.62.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.2.6
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pillow==12.1.1
protobuf==7.34.1
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
tensorflow==2.21.0
termcolor==3.3.0
tomli==2.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
